Priority + reminder_lead_minutes: BIGINT cast for dynamic update path
- action.priority BIGINT, but update handler binds as text. Add explicit ${n}::BIGINT cast in dynamic SET clause.
- Same fix applied to event.reminder_lead_minutes for symmetry.
- All endpoints verified end-to-end: register/login/me/refresh, contacts+tag_ids,
  actions CRUD, events, projects, reminders, interactions, sync manifest/pull.

+++ server/src/handlers/event.rs
@@ -92,7 +92,7 @@ pub async fn create(
     .bind(body.get("notes").and_then(|v| v.as_str()))
     .bind(body.get("contact_id").and_then(|v| v.as_str()))
     .bind(body.get("project_id").and_then(|v| v.as_str()))
-    .bind(body.get("reminder_lead_minutes").and_then(|v| v.as_i64()).map(|n| n as i32))
+    .bind(body.get("reminder_lead_minutes").and_then(|v| v.as_i64()).map(|n| n as i64))
     .bind(&now)
     .bind(&now)
     .execute(&mut *tx)
